Rename lfpackage to language
