+++ ledger/participant-state-index/src/main/scala/com/daml/ledger/participant/state/index/v2/IndexPackagesService.scala
@@ -5,7 +5,7 @@ package com.daml.ledger.participant.state.index.v2
 
 import com.digitalasset.daml.lf.data.Ref.PackageId
 import com.digitalasset.daml_lf.DamlLf.Archive
-import com.digitalasset.daml.lf.lfpackage.Ast.Package
+import com.digitalasset.daml.lf.language.Ast.Package
 
 import scala.concurrent.Future
